Move to asynchronous command processing
In order to handle command line input simultaneously with network traffic,
I built a message loop that checks for input from all sources, and creates
an event in the message queue when input is received. Then the messages in
the queue are handled after all input has been checked, ensuring ordering
and responsiveness.

Signed-off-by: Joel Nider <joel@nider.org>